requirements.txt
asgiref==3.8.1
Django==5.0.3
django-ckeditor==6.7.1
django-js-asset==2.2.0
djangorestframework==3.15.1
Markdown==3.6
mysqlclient==2.2.4
pillow==10.2.0
python-decouple==3.8
sqlparse==0.4.4
tzdata==2024.1
